# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationDocumentResponse": {
        "properties": {
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "material_revision": {
            "title": "Material Revision",
            "type": "integer"
          },
          "media_type": {
            "enum": [
              "application/pdf",
              "text/plain",
              "text/markdown"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "id",
          "material_id",
          "material_revision",
          "original_filename",
          "media_type",
          "byte_size",
          "sha256",
          "created_at"
        ],
        "title": "ApplicationDocumentResponse",
        "type": "object"
      },
...
      "Body_attach_document_api_application_materials__material_id__revisions__material_revision__document_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_attach_document_api_application_materials__material_id__revisions__material_revision__document_post",
        "type": "object"
      },
...
      "DuplicateCaseReviewResponse": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DuplicateDecisionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision_history": {
            "items": {
              "$ref": "#/components/schemas/DuplicateDecisionResponse"
            },
            "title": "Decision History",
            "type": "array"
          },
          "evidence_summary": {
            "title": "Evidence Summary",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_resolved": {
            "title": "Is Resolved",
            "type": "boolean"
          },
          "is_reviewed": {
            "title": "Is Reviewed",
            "type": "boolean"
          },
          "left_subject": {
            "$ref": "#/components/schemas/DuplicateSubjectSummaryResponse"
          },
          "right_subject": {
            "$ref": "#/components/schemas/DuplicateSubjectSummaryResponse"
          },
          "source_references": {
            "items": {
              "$ref": "#/components/schemas/DuplicateSourceReferenceSummaryResponse"
            },
            "title": "Source References",
            "type": "array"
          },
          "subject_type": {
            "enum": [
              "opportunity",
              "posting"
            ],
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subject_type",
          "left_subject",
          "right_subject",
          "evidence_summary",
          "confidence",
          "source_references",
          "created_at",
          "current_decision",
          "decision_history",
          "is_reviewed",
          "is_resolved"
        ],
        "title": "DuplicateCaseReviewResponse",
        "type": "object"
      },
      "DuplicateDecisionPayload": {
        "additionalProperties": false,
        "properties": {
          "outcome": {
            "enum": [
              "confirmed_duplicate",
              "confirmed_distinct",
              "related_but_distinct",
              "keep_unresolved"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "outcome",
          "reason"
        ],
        "title": "DuplicateDecisionPayload",
        "type": "object"
      },
...
      "ExternalLinkOpenPayload": {
        "additionalProperties": false,
        "properties": {
          "posting_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Id"
          }
        },
        "title": "ExternalLinkOpenPayload",
        "type": "object"
      },
      "ExternalLinkResponse": {
        "properties": {
          "availability": {
            "enum": [
              "available",
              "unavailable",
              "uncertain",
              "unknown"
            ],
            "title": "Availability",
            "type": "string"
          },
          "display_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Label"
          },
          "observed_at": {
            "title": "Observed At",
            "type": "string"
          },
          "posting_id": {
            "title": "Posting Id",
            "type": "string"
          },
          "preferred": {
            "title": "Preferred",
            "type": "boolean"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "company_careers",
              "job_board",
              "professional_network",
              "other"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "posting_id",
          "source_id",
          "source_name",
          "source_type",
          "url",
          "display_label",
          "availability",
          "observed_at",
          "preferred"
        ],
        "title": "ExternalLinkResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpportunityDetailResponse": {
        "properties": {
          "assessments": {
            "items": {
              "$ref": "#/components/schemas/AssessmentResponse"
            },
            "title": "Assessments",
            "type": "array"
          },
          "availability": {
            "anyOf": [
              {
                "enum": [
                  "available",
                  "unavailable",
                  "uncertain",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability"
          },
          "availability_age_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Age Days"
          },
          "availability_last_checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Last Checked At"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyResponse"
          },
          "decision_history": {
            "items": {
              "$ref": "#/components/schemas/DecisionResponse"
            },
            "title": "Decision History",
            "type": "array"
          },
          "external_assessments": {
            "items": {
              "$ref": "#/components/schemas/AssessmentResponse"
            },
            "title": "External Assessments",
            "type": "array"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/OpportunityGroupSummaryResponse"
            },
            "title": "Groups",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_provenance": {
            "$ref": "#/components/schemas/ImportProvenanceResponse"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationResponse"
            },
            "title": "Locations",
            "type": "array"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/ObservationResponse"
            },
            "title": "Observations",
            "type": "array"
          },
          "personal_assessment_history": {
            "items": {
              "$ref": "#/components/schemas/PersonalAssessmentResponse"
            },
            "title": "Personal Assessment History",
            "type": "array"
          },
          "personal_assessments": {
            "items": {
              "$ref": "#/components/schemas/PersonalAssessmentResponse"
            },
            "title": "Personal Assessments",
            "type": "array"
          },
          "postings": {
            "items": {
              "$ref": "#/components/schemas/PostingResponse"
            },
            "title": "Postings",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceResponse"
            },
            "title": "Sources",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tracking_status": {
            "enum": [
              "new",
              "to_review",
              "interesting",
              "shortlisted",
              "deferred",
              "excluded",
              "archived"
            ],
            "title": "Tracking Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "locations",
          "postings",
          "sources",
          "observations",
          "tracking_status",
          "external_assessments",
          "assessments",
          "personal_assessments",
          "personal_assessment_history",
          "decision_history",
          "import_provenance"
        ],
        "title": "OpportunityDetailResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vocation",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/application-documents/{document_id}": {
      "get": {
        "operationId": "get_document_api_application_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document",
        "tags": [
          "application-documents"
        ]
      }
    },
    "/api/application-documents/{document_id}/content": {
      "get": {
        "operationId": "get_content_api_application_documents__document_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Content",
        "tags": [
          "application-documents"
        ]
      }
    },
...
    "/api/application-materials/{material_id}/revisions/{material_revision}/document": {
      "get": {
        "operationId": "get_for_material_revision_api_application_materials__material_id__revisions__material_revision__document_get",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "material_revision",
            "required": true,
            "schema": {
              "title": "Material Revision",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get For Material Revision",
        "tags": [
          "application-documents"
        ]
      },
      "post": {
        "operationId": "attach_document_api_application_materials__material_id__revisions__material_revision__document_post",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "material_revision",
            "required": true,
            "schema": {
              "title": "Material Revision",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_attach_document_api_application_materials__material_id__revisions__material_revision__document_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Document",
        "tags": [
          "application-documents"
        ]
      }
    },
...
    "/api/duplicate-cases": {
      "get": {
        "operationId": "list_duplicate_cases_api_duplicate_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "subject_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "opportunity",
                    "posting"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Type"
            }
          },
          {
            "in": "query",
            "name": "subject_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DuplicateCaseReviewResponse"
                  },
                  "title": "Response List Duplicate Cases Api Duplicate Cases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Duplicate Cases",
        "tags": [
          "duplicate-cases"
        ]
      }
    },
    "/api/duplicate-cases/{case_id}": {
      "get": {
        "operationId": "get_duplicate_case_api_duplicate_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateCaseReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Duplicate Case",
        "tags": [
          "duplicate-cases"
        ]
      }
    },
    "/api/duplicate-cases/{case_id}/decisions": {
      "post": {
        "operationId": "decide_duplicate_case_api_duplicate_cases__case_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateDecisionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateCaseReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Duplicate Case",
        "tags": [
          "duplicate-cases"
        ]
      }
    },
    "/api/external-links/opportunities/{opportunity_id}": {
      "get": {
        "operationId": "opportunity_links_api_external_links_opportunities__opportunity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExternalLinkResponse"
                  },
                  "title": "Response Opportunity Links Api External Links Opportunities  Opportunity Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Links",
        "tags": [
          "external-links"
        ]
      }
    },
    "/api/external-links/opportunities/{opportunity_id}/open": {
      "post": {
        "operationId": "open_opportunity_api_external_links_opportunities__opportunity_id__open_post",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalLinkOpenPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalLinkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Opportunity",
        "tags": [
          "external-links"
        ]
      }
    },
...
    "/api/opportunities/{opportunity_id}": {
      "get": {
        "operationId": "opportunity_detail_api_opportunities__opportunity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Detail",
        "tags": [
          "opportunities"
        ]
      }
    },
...
  }
}